Time: 5 ms (32.52%), Space: 69.6 MB (44.54%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,6 +1,6 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        int currentsum = 0;
+        double currentsum = 0;
         for(int i=0;i<k;i++){
             currentsum += nums[i];
